seperated gen token function to new generate-token file and imported into log in controllers
